requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow>=9.5.0
gunicorn>=21.2.0


